This commit contains the translation implementation and the third button with the variation on android Q.
Translations are done by creating values for English (values-en) and values for Portuguese (values-pt). Within the XML it contains the string variables with the native text of each language.

Different text and color on Android >= Q is possible with the creation of de values to specific version of android (values-v29).